CRUD Model_Modele : Desactivation d'un  model, mais le model reste dans la BD

+++ model/Model_Modele.php
@@ -1,15 +1,17 @@
 <?php
+	require_once("TemplateDAO.php");  //Ajouter pour le test de la class
 	class Model_Modele extends TemplateDAO {
 		
 		public function getTable() {
 			return "modele";
 		}
-
+		
 		/* Differentes methodes CRUD modele   */
-		public function obtenirListeModele() { 
+		
+		public function obtenirListeModele() {
 			try {
-				$stmt = $this->connexion->query("SELECT m.idModele as idModele, m.modele as modele, ma.marque as marque from modele m JOIN marque ma on m.idMarque = ma.idMarque");
-
+				//$stmt = $this->connexion->query("SELECT idModele, idMarque,modele FROM modele where status=1");
+				$stmt = $this->connexion->query("SELECT m.idModele as idModele, m.modele as modele, ma.marque as marque from modele m JOIN marque ma on m.idMarque = ma.idMarque where m.status=1");
 				$stmt->execute();
 				return $stmt->fetchAll();
 
@@ -18,6 +20,7 @@
 				return 0;
 			}
 		}
+
 		public function obtenirFabricantModel($idModele){
 			try {
 				$stmt = $this->connexion->query("SELECT fabricant, idModele FROM modele 
@@ -34,13 +37,17 @@
 			}
 		}
 
+
+
+
+
 		//Chercher  modèle par idMarque
 		public function getListeModeleByMarque($idMarque) {
 			try {
-							$stmt = $this->connexion->query("SELECT idModele, idMarque,modele FROM modele where idMarque='" . $idMarque."'");
+							$stmt = $this->connexion->query("SELECT idModele, idMarque,modele FROM modele where status=1  && idMarque='" . $idMarque."'");
 							$stmt->execute();
 							return $stmt->fetchAll();
-							//return 1; //L'utilisateur existe dans la BD
+							//return 1;
 			
 			}
 						catch(Exception $exc) {
@@ -67,14 +74,26 @@
 				// Supprimer modele dans la BD
 		public function supprimerModele($id) {		
 			try {
-				$stmt = $this->connexion->prepare("DELETE FROM modele WHERE idModele=" . $id);
-				$stmt->execute();
-				return $stmt->rowCount();
+				//Vérifier si modèle existe dans la table voiture
+				$existe = $this->chercherModeledasnVoiture($id) ;         
+				if ($existe["cont"]  >= 1 )
+				{					
+					$stmt = $this->connexion->prepare("UPDATE modele SET status=0 WHERE idModele=".$id);
+					$stmt->execute();
+				}
+				else
+				{	
+						$stmt = $this->connexion->prepare("DELETE FROM modele WHERE idModele=" . $id);
+						$stmt->execute();
+				}
+				//return $stmt->rowCount();
+				return 1;
 
 			}
 			catch(Exception $exc) {
 				return 0;
 			}
+
 		}
 
 		// Modifier un modèle dans la BD
@@ -91,15 +110,22 @@
 			catch(Exception $exc) {
 				return 0;
 			}
-		}		
+		}
 
-
-
-
-
-
-
-
+				//Chercher  modele dans la table voiture
+				public function chercherModeledasnVoiture($idModele) {
+					try {
+									$stmt = $this->connexion->query("SELECT count(*) cont FROM voiture where idModele='" . $idModele."'");
+									$stmt->execute();
+									//return 1;
+									return $stmt->fetch();
+									//return 1; //L'utilisateur existe dans la BD
+					
+					}
+								catch(Exception $exc) {
+									return 0;
+								}
+					}
 
 	}
 ?>
